Update cache-invalidation.md
Some resources like datasources and tags have some issues that this document does not resolve:
- they do not send a `cv` parameter in their response
- Data sent from these endpoints is often stale

The way to receive the most up-to-date version of data from these endpoints is to send a `cv` parameter set to the current timestamp. I've made some edits to the doc both to clarify the role of the `cv` parameter and to explain how to use it in cases when the response does not send back a `cv` param.